requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.104.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
jiter==0.11.1
openai==2.8.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.32.5
sarvamai
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
